Split scenarios into separate files
Our testing style guide is one system spec per file
