+++ spec/system/provider_interface/see_applications_spec.rb
@@ -4,17 +4,6 @@ RSpec.feature 'See applications' do
   include CourseOptionHelpers
   include DfESignInHelpers
 
-  scenario 'Provider visits applications when there are none' do
-    given_i_am_a_provider_user_authenticated_with_dfe_sign_in
-    and_my_training_provider_exists
-    and_another_organisation_has_applications
-
-    when_i_have_been_assigned_to_my_training_provider
-    and_i_sign_in_to_the_provider_interface
-
-    then_i_should_see_no_applications
-  end
-
   scenario 'Provider visits application page' do
     given_i_am_a_provider_user_authenticated_with_dfe_sign_in
     and_my_organisation_has_applications
@@ -59,12 +48,6 @@ RSpec.feature 'See applications' do
     @my_provider_choice2  = create(:submitted_application_choice, status: 'awaiting_provider_decision', course_option: course_option)
   end
 
-  def and_another_organisation_has_applications
-    other_course_option = course_option_for_provider_code(provider_code: 'ANOTHER_ORG')
-
-    @other_provider_choice = create(:submitted_application_choice, status: 'awaiting_provider_decision', course_option: other_course_option)
-  end
-
   def and_i_visit_the_provider_page
     visit provider_interface_path
   end
@@ -76,15 +59,6 @@ RSpec.feature 'See applications' do
     expect(page).to have_content @my_provider_choice1.application_form.first_name
     expect(page).to have_content @my_provider_choice2.application_form.support_reference
     expect(page).to have_content @my_provider_choice2.application_form.first_name
-  end
-
-  def then_i_should_see_no_applications
-    expect(page).to have_content 'You haven’t received any applications'
-    expect(page).not_to have_selector('.govuk-table')
-  end
-
-  def but_not_the_applications_from_other_providers
-    expect(page).not_to have_content @other_provider_choice.application_form.first_name
   end
 
   def when_i_click_on_an_application
